Move exceptions into their own packages
