+++ src/main/kotlin/no/nb/mlt/wls/domain/WLSService.kt
@@ -27,10 +27,8 @@ import no.nb.mlt.wls.domain.ports.inbound.GetItems
 import no.nb.mlt.wls.domain.ports.inbound.GetOrder
 import no.nb.mlt.wls.domain.ports.inbound.GetOrders
 import no.nb.mlt.wls.domain.ports.inbound.ItemMetadata
-import no.nb.mlt.wls.domain.ports.inbound.ItemNotFoundException
 import no.nb.mlt.wls.domain.ports.inbound.MoveItem
 import no.nb.mlt.wls.domain.ports.inbound.MoveItemPayload
-import no.nb.mlt.wls.domain.ports.inbound.OrderNotFoundException
 import no.nb.mlt.wls.domain.ports.inbound.PickItems
 import no.nb.mlt.wls.domain.ports.inbound.PickOrderItems
 import no.nb.mlt.wls.domain.ports.inbound.StockCount
@@ -38,6 +36,8 @@ import no.nb.mlt.wls.domain.ports.inbound.SynchronizeItems
 import no.nb.mlt.wls.domain.ports.inbound.UpdateItem
 import no.nb.mlt.wls.domain.ports.inbound.UpdateItem.UpdateItemPayload
 import no.nb.mlt.wls.domain.ports.inbound.UpdateOrderStatus
+import no.nb.mlt.wls.domain.ports.inbound.exceptions.ItemNotFoundException
+import no.nb.mlt.wls.domain.ports.inbound.exceptions.OrderNotFoundException
 import no.nb.mlt.wls.domain.ports.outbound.DuplicateResourceException
 import no.nb.mlt.wls.domain.ports.outbound.EventProcessor
 import no.nb.mlt.wls.domain.ports.outbound.EventRepository
@@ -369,7 +369,7 @@ class WLSService(
         val orders: List<Order> = orderRepository.getOrdersWithItems(hostName, returnedItems)
 
         orders.forEach { order ->
-            val (_, orderEvent) =
+            val (returnedOrder, orderEvent) =
                 transactionPort.executeInTransaction {
                     val returnOrder = order.returnItems(returnedItems)
                     val updatedOrder = orderRepository.updateOrder(returnOrder)
@@ -378,7 +378,9 @@ class WLSService(
                 } ?: (order to null)
 
             if (orderEvent == null) {
-                logger.error { "Failed to properly mark order: ${order.hostOrderId} in host: ${order.hostName} as returned, transaction failed" }
+                logger.error {
+                    "Failed to properly mark order: ${returnedOrder.hostOrderId} in host: ${returnedOrder.hostName} as returned, transaction failed"
+                }
                 return@forEach
             }
