+++ src/main/kotlin/no/nb/mlt/wls/domain/ports/inbound/GetOrder.kt
@@ -2,6 +2,7 @@ package no.nb.mlt.wls.domain.ports.inbound
 
 import no.nb.mlt.wls.domain.model.HostName
 import no.nb.mlt.wls.domain.model.Order
+import no.nb.mlt.wls.domain.ports.inbound.exceptions.OrderNotFoundException
 
 /**
  * A port for retrieving an order based on a given host name and order ID.
