+++ src/main/kotlin/no/nb/mlt/wls/domain/ports/inbound/exceptions/ItemNotFoundException.kt
@@ -1,4 +1,4 @@
-package no.nb.mlt.wls.domain.ports.inbound
+package no.nb.mlt.wls.domain.ports.inbound.exceptions
 
 class ItemNotFoundException(
     override val message: String
