+++ src/main/kotlin/no/nb/mlt/wls/domain/ports/inbound/exceptions/OrderNotFoundException.kt
@@ -1,4 +1,4 @@
-package no.nb.mlt.wls.domain.ports.inbound
+package no.nb.mlt.wls.domain.ports.inbound.exceptions
 
 class OrderNotFoundException(
     override val message: String
