+++ src/main/kotlin/no/nb/mlt/wls/domain/ports/inbound/exceptions/ServerException.kt
@@ -1,4 +1,4 @@
-package no.nb.mlt.wls.domain.ports.inbound
+package no.nb.mlt.wls.domain.ports.inbound.exceptions
 
 class ServerException(
     override val message: String,
