+++ src/main/kotlin/no/nb/mlt/wls/domain/ports/inbound/exceptions/ValidationException.kt
@@ -1,4 +1,4 @@
-package no.nb.mlt.wls.domain.ports.inbound
+package no.nb.mlt.wls.domain.ports.inbound.exceptions
 
 class ValidationException(
     override val message: String,
