+++ src/test/kotlin/no/nb/mlt/wls/domain/model/OrderTest.kt
@@ -1,7 +1,7 @@
 package no.nb.mlt.wls.domain.model
 
 import no.nb.mlt.wls.createTestOrder
-import no.nb.mlt.wls.domain.ports.inbound.IllegalOrderStateException
+import no.nb.mlt.wls.domain.ports.inbound.exceptions.IllegalOrderStateException
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Assertions.assertThrows
 import org.junit.jupiter.api.Test
