+++ src/test/kotlin/no/nb/mlt/wls/item/model/ItemModelValidationTest.kt
@@ -3,7 +3,7 @@ package no.nb.mlt.wls.item.model
 import no.nb.mlt.wls.application.hostapi.item.toApiPayload
 import no.nb.mlt.wls.createTestItem
 import no.nb.mlt.wls.domain.model.UNKNOWN_LOCATION
-import no.nb.mlt.wls.domain.ports.inbound.ValidationException
+import no.nb.mlt.wls.domain.ports.inbound.exceptions.ValidationException
 import org.assertj.core.api.Assertions.catchThrowable
 import org.assertj.core.api.BDDAssertions.then
 import org.assertj.core.api.BDDAssertions.thenCode
